# api/swagger/swagger.yaml
basePath: /api/1.0/
definitions:
...
  github_com_openinfradev_tks-api_pkg_domain.AddOrganizationSystemNotificationTemplatesRequest:
    properties:
      systemNotificationTemplateIds:
        items:
          type: string
        type: array
    required:
    - systemNotificationTemplateIds
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.CheckSystemNotificaionTemplateNameResponse:
    properties:
      existed:
        type: boolean
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.CreateSystemNotificationTemplateRequest:
    properties:
      description:
        type: string
      metricParameters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationMetricParameterResponse'
        type: array
      metricQuery:
        type: string
      name:
        type: string
      organizationIds:
        items:
          type: string
        type: array
    required:
    - metricQuery
    - name
    - organizationIds
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.FilterResponse:
    properties:
      column:
        type: string
      operator:
        type: string
      releation:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationTemplateResponse:
    properties:
      systemNotificationTemplate:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationTemplateResponse'
    type: object
  github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationTemplatesResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.PaginationResponse'
      systemNotificationTemplates:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationTemplateResponse'
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.PaginationResponse:
    properties:
      filters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FilterResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sortColumn:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.RemoveOrganizationSystemNotificationTemplatesRequest:
    properties:
      systemNotificationTemplateIds:
        items:
          type: string
        type: array
    required:
    - systemNotificationTemplateIds
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleOrganizationResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse:
    properties:
      accountId:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SystemNotificationMetricParameterResponse:
    properties:
      key:
        type: string
      order:
        type: integer
      value:
        type: string
    required:
    - key
    - order
    - value
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SystemNotificationTemplateResponse:
    properties:
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
      description:
        type: string
      id:
        type: string
      isSystem:
        type: boolean
      metricParameters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationMetricParameterResponse'
        type: array
      metricQuery:
        type: string
      name:
        type: string
      notificationType:
        type: string
      organizations:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleOrganizationResponse'
        type: array
      updatedAt:
        type: string
      updator:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
    required:
    - metricQuery
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.UpdateSystemNotificationTemplateRequest:
    properties:
      description:
        type: string
      metricParameters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SystemNotificationMetricParameterResponse'
        type: array
      metricQuery:
        type: string
      name:
        type: string
      organizationIds:
        items:
          type: string
        type: array
    required:
    - metricQuery
    - name
    - organizationIds
    type: object
...
host: tks-api-dev.taco-cat.xyz
info:
  contact:
    email: taekyu.kang@sk.com
    name: taekyu.kang@sk.com
  description: This is backend api service for tks platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: tks-api service
  version: "1.0"
paths:
...
  /admin/system-notification-templates:
    get:
      consumes:
      - application/json
      description: Get SystemNotificationTemplates
      parameters:
      - description: pageSize
        in: query
        name: pageSize
        type: string
      - description: pageNumber
        in: query
        name: pageNumber
        type: string
      - description: sortColumn
        in: query
        name: soertColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - collectionFormat: csv
        description: filters
        in: query
        items:
          type: string
        name: filters
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationTemplatesResponse'
      security:
      - JWT: []
      summary: Get SystemNotificationTemplates
      tags:
      - SystemNotificationTemplates
    post:
      consumes:
      - application/json
      description: Create alert template. ADMIN ONLY
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.CreateSystemNotificationTemplateRequest'
      security:
      - JWT: []
      summary: Create alert template. ADMIN ONLY
      tags:
      - SystemNotificationTemplates
  /admin/system-notification-templates/{systemNotificationTemplateId}:
    delete:
      consumes:
      - application/json
      description: Delete SystemNotificationTemplate
      parameters:
      - description: systemNotificationTemplateId
        in: path
        name: systemNotificationTemplateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Delete SystemNotificationTemplate
      tags:
      - SystemNotificationTemplates
    get:
      consumes:
      - application/json
      description: Get SystemNotificationTemplate
      parameters:
      - description: systemNotificationTemplateId
        in: path
        name: systemNotificationTemplateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationTemplateResponse'
      security:
      - JWT: []
      summary: Get SystemNotificationTemplate
      tags:
      - SystemNotificationTemplates
    put:
      consumes:
      - application/json
      description: Update SystemNotificationTemplate
      parameters:
      - description: systemNotificationTemplateId
        in: path
        name: systemNotificationTemplateId
        required: true
        type: string
      - description: Update alert template request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.UpdateSystemNotificationTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Update SystemNotificationTemplate
      tags:
      - SystemNotificationTemplates
  /admin/system-notification-templates/name/{name}/existence:
    get:
      consumes:
      - application/json
      description: Check name for systemNotificationTemplate
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.CheckSystemNotificaionTemplateNameResponse'
      security:
      - JWT: []
      summary: Check name for systemNotificationTemplate
      tags:
      - SystemNotificationTemplates
...
  /organizations/{organizationId}/system-notification-templates:
    get:
      consumes:
      - application/json
      description: Get Organization SystemNotificationTemplates
      parameters:
      - description: pageSize
        in: query
        name: pageSize
        type: string
      - description: pageNumber
        in: query
        name: pageNumber
        type: string
      - description: sortColumn
        in: query
        name: soertColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - collectionFormat: csv
        description: filters
        in: query
        items:
          type: string
        name: filters
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationTemplatesResponse'
      security:
      - JWT: []
      summary: Get Organization SystemNotificationTemplates
      tags:
      - SystemNotificationTemplates
    post:
      consumes:
      - application/json
      description: Add organization systemNotificationTemplates
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: Add organization systemNotification templates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.AddOrganizationSystemNotificationTemplatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Add organization systemNotificationTemplates
      tags:
      - SystemNotificationTemplates
    put:
      consumes:
      - application/json
      description: Remove organization systemNotificationTemplates
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: Remove organization systemNotification templates request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.RemoveOrganizationSystemNotificationTemplatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - JWT: []
      summary: Remove organization systemNotificationTemplates
      tags:
      - SystemNotificationTemplates
  /organizations/{organizationId}/system-notification-templates/{systemNotificationTemplateId}:
    get:
      consumes:
      - application/json
      description: Get Organization SystemNotificationTemplate
      parameters:
      - description: systemNotificationTemplateId
        in: path
        name: systemNotificationTemplateId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetSystemNotificationTemplateResponse'
      security:
      - JWT: []
      summary: Get Organization SystemNotificationTemplate
      tags:
      - SystemNotificationTemplates
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"